+++ test/ssh/test_vulnerabilities.py
@@ -4,10 +4,12 @@ from cryptodatahub.ssh.algorithm import SshKexAlgorithm, SshEncryptionAlgorithm,
 
 from cryptoparser.common.base import Serializable, SerializableTextEncoder
 
+from cryptolyzer.common.exception import NetworkError
 from cryptolyzer.common.transfer import L4TransferSocketParams
 from cryptolyzer.common.utils import SerializableTextEncoderHighlighted
 
-from cryptolyzer.ssh.client import L7ClientSsh
+from cryptolyzer.ssh.client import L7ClientSsh, SshDisconnect
+from cryptolyzer.ssh.server import L7ServerSsh, SshServerConfiguration
 from cryptolyzer.ssh.vulnerabilities import (
     AnalyzerResultVulnerabilities,
     AnalyzerResultVulnerabilityAlgorithms,
@@ -16,7 +18,7 @@ from cryptolyzer.ssh.vulnerabilities import (
     AnalyzerVulnerabilities,
 )
 
-from .classes import TestSshCases
+from .classes import L7ServerSshTest, TestSshCases
 
 
 class TestSshVulnerabilities(TestSshCases.TestSshClientBase):
@@ -177,3 +179,16 @@ class TestSshVulnerabilities(TestSshCases.TestSshClientBase):
 
         log_stream = '\n'.join(self.pop_log_lines())
         self._check_gex_params([2048, 3072, 4096, 6144, 8192], log_stream)
+
+    def test_vulns_with_algorithm_limit(self):
+        server_configuration = SshServerConfiguration(max_remote_algorithm_count=50)
+        threaded_server = L7ServerSshTest(L7ServerSsh(
+            'localhost', 0, L4TransferSocketParams(timeout=0.2), configuration=server_configuration
+        ))
+        threaded_server.start()
+
+        try:
+            result = self.get_result('localhost', threaded_server.l7_server.l4_transfer.bind_port)
+            self.assertIsNotNone(result)
+        except (NetworkError, SshDisconnect, StopIteration):
+            pass